Rename to parse friendly player
