+++ app/scripts/parse-friendly-player.js
